+++ src/components/skeleton/InfoCardSkeleton.tsx
